Fix fetch bug
Because mutliprocessing runs in separate processes, it is necessary
to re-load podcast details between update and download (as the object
in memory won't be updated because the updated object is only
updated in a different process' memory.

Also pass args through to update and download so we can have behaviour
based on the args.verbose option.